# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App core",
        "version": "0.1"
    },
    "paths": {
        "/areas/{area_id}": {
            "get": {
                "tags": [
                    "areas"
                ],
                "summary": "Get area from id",
                "operationId": "App\\Http\\Controllers\\Areas\\AreaShowController::__invoke",
                "parameters": [
                    {
                        "name": "area_id",
                        "in": "path",
                        "description": "area id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/AreaResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não foi possível encontrar a área"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "areas"
                ],
                "summary": "Update area",
                "operationId": "App\\Http\\Controllers\\Areas\\AreaUpdateController::__invoke",
                "parameters": [
                    {
                        "name": "area_id",
                        "in": "path",
                        "description": "area id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create area",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAreaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/AreaResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não foi possível atualizar a área"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "areas"
                ],
                "summary": "Delete area",
                "operationId": "App\\Http\\Controllers\\Areas\\AreaDestroyController::__invoke",
                "parameters": [
                    {
                        "name": "area_id",
                        "in": "path",
                        "description": "area id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "Não foi possível excluir a área"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/areas": {
            "get": {
                "tags": [
                    "areas"
                ],
                "summary": "List of areas",
                "operationId": "App\\Http\\Controllers\\Areas\\AreaIndexController::__invoke",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AreaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "areas"
                ],
                "summary": "Create area",
                "operationId": "App\\Http\\Controllers\\Areas\\AreaStoreController::__invoke",
                "requestBody": {
                    "description": "Create area",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAreaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/AreaResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Falha ao criar área"
                    },
                    "422": {
                        "description": "Informa os campos inválidos ou faltando"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/auth/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Auth me",
                "operationId": "App\\Http\\Controllers\\Auth\\MeShowController::__invoke",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "auth"
                ],
                "summary": "Update User Auth me",
                "operationId": "App\\Http\\Controllers\\Auth\\MeUpdateController::__invoke",
                "requestBody": {
                    "description": "Update User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Falha na atualização do usuário"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/auth/password": {
            "put": {
                "tags": [
                    "auth"
                ],
                "summary": "Update Password to User",
                "operationId": "App\\Http\\Controllers\\Auth\\MeUpdatePasswordController::__invoke",
                "requestBody": {
                    "description": "Update password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/MessageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbiden"
                    },
                    "500": {
                        "description": "Falha na atualização do usuário"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/goals/{goal_id}": {
            "get": {
                "tags": [
                    "goals"
                ],
                "summary": "Get goal from id",
                "operationId": "App\\Http\\Controllers\\Goals\\GoalShowController::__invoke",
                "parameters": [
                    {
                        "name": "goal_id",
                        "in": "path",
                        "description": "goal id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/GoalResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não foi possível encontrar o objetivo"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "goals"
                ],
                "summary": "Update goal",
                "operationId": "App\\Http\\Controllers\\Goals\\GoalUpdateController::__invoke",
                "parameters": [
                    {
                        "name": "goal_id",
                        "in": "path",
                        "description": "goal id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create goal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGoalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/GoalResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não foi possível atualizar o objetivo"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "goals"
                ],
                "summary": "Delete goal",
                "operationId": "App\\Http\\Controllers\\Goals\\GoalDestroyController::__invoke",
                "parameters": [
                    {
                        "name": "goal_id",
                        "in": "path",
                        "description": "goal id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "Não foi possível excluir o objetivo"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/goals": {
            "get": {
                "tags": [
                    "goals"
                ],
                "summary": "List of goals",
                "operationId": "App\\Http\\Controllers\\Goals\\GoalIndexController::__invoke",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GoalResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "goals"
                ],
                "summary": "Create goal",
                "operationId": "App\\Http\\Controllers\\Goals\\GoalStoreController::__invoke",
                "requestBody": {
                    "description": "Create goal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGoalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/GoalResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Falha ao criar objetivo"
                    },
                    "422": {
                        "description": "Informa os campos inválidos ou faltando"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/gratitude-diaries/{gratitude_diary_id}": {
            "get": {
                "tags": [
                    "gratitude-diaries"
                ],
                "summary": "Get gratitude_diary from id",
                "operationId": "App\\Http\\Controllers\\GratitudeDiaries\\GratitudeDiaryShowController::__invoke",
                "parameters": [
                    {
                        "name": "gratitudediary_id",
                        "in": "path",
                        "description": "gratitude diary id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/GratitudeDiaryResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não foi possível encontrar a gratidão diária"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "gratitude-diaries"
                ],
                "summary": "Update gratitude_diary",
                "operationId": "App\\Http\\Controllers\\GratitudeDiaries\\GratitudeDiaryUpdateController::__invoke",
                "parameters": [
                    {
                        "name": "gratitude_diary_id",
                        "in": "path",
                        "description": "gratitude_diary id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create gratitude diary",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGratitudeDiaryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/GratitudeDiaryResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não foi possível atualizar a gratidão diária"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "gratitude-diaries"
                ],
                "summary": "Delete gratitude_diary",
                "operationId": "App\\Http\\Controllers\\GratitudeDiaries\\GratitudeDiaryDestroyController::__invoke",
                "parameters": [
                    {
                        "name": "gratitude_diary_id",
                        "in": "path",
                        "description": "gratitude_diary id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "Não foi possível excluir a gratidão diária"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/gratitude-diaries": {
            "get": {
                "tags": [
                    "gratitude-diaries"
                ],
                "summary": "List of gratitude diaries",
                "operationId": "App\\Http\\Controllers\\GratitudeDiaries\\GratitudeDiaryIndexController::__invoke",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GratitudeDiaryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "gratitude-diaries"
                ],
                "summary": "Create gratitude diary",
                "operationId": "App\\Http\\Controllers\\GratitudeDiaries\\GratitudeDiaryStoreController::__invoke",
                "requestBody": {
                    "description": "Create gratitude diary",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGratitudeDiaryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/GratitudeDiaryResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Falha ao criar gratidão diária"
                    },
                    "422": {
                        "description": "Informa os campos inválidos ou faltando"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/": {
            "get": {
                "tags": [
                    "index"
                ],
                "summary": "Index",
                "operationId": "App\\Http\\Controllers\\IndexController::__invoke",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/IndexResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateAreaRequest": {
                "required": [
                    "name",
                    "icon"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "comments": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateAreaRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "comments": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdatePasswordRequest": {
                "properties": {
                    "current_password": {
                        "description": "Senha atual",
                        "type": "string",
                        "format": "password"
                    },
                    "password": {
                        "description": "Nova senha",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
...
            "UserUpdateRequest": {
                "properties": {
                    "name": {
                        "description": "Nome",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "AreaResource": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "comments": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
            "GoalResource": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "area_id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "comments": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "MessageResource": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    }
}